+++ src/types.ts
@@ -87,42 +87,3 @@ export type RequestExtendedOptions<V = Variables> = { url: string } & RequestOpt
 export type RawRequestExtendedOptions<V = Variables> = { url: string } & RawRequestOptions<V>
 
 export type BatchRequestsExtendedOptions<V = Variables> = { url: string } & BatchRequestsOptions<V>
-
-export interface GraphQLRequestClient {
-  rawRequest<T = any, V = Variables, E = any>(
-    query: string,
-    variables?: V,
-    requestHeaders?: Dom.RequestInit['headers']
-  ): Promise<{ data: T; extensions?: E; headers?: Dom.Headers; status?: number }>
-
-  request<T = any, V = Variables>(
-    document: RequestDocument,
-    variables?: V,
-    requestHeaders?: Dom.RequestInit['headers']
-  ): Promise<T>
-
-}
-
-export type UnsubscribeCallback = () => void;
-
-export interface GraphQLSubscriber<T, E=unknown> {
-  next?(data: T, extensions?: E): void;
-  error?(errorValue: ClientError): void;
-  complete?(): void;
-}
-
-export interface GraphQLSubscriptionClient extends GraphQLRequestClient {
-  subscribe<T = any, V = Variables, E = any>(
-      document: RequestDocument,
-      subscriber: GraphQLSubscriber<T, E>,
-      variables?: V
-  ): UnsubscribeCallback;
-
-  rawSubscribe<T = any, V = Variables, E = any>(
-      query: string,
-      subscriber: GraphQLSubscriber<T, E>,
-      variables?: V
-  ): UnsubscribeCallback;
-
-  ping(payload: Variables): void;
-}
